requirements.txt
# requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.9
numpy>=1.19.0
pyyaml>=5.4.1